Dynamic RPM on chart
Added dynamic chart update of rpm

+++ Hardware/temp_acquisition/temp_acquisition.ino
@@ -22,7 +22,7 @@ float min_temp = 76;              // Minimum temperature that will represent 0% 
 float temp_range;                 // Variable used to calculate the range in which pwm duty cycle % will be mappped to.
 const int pwm_max_bits = 255;     // 255=100% : 0=0% duty cycle for Arduino analogWrite() API call
 int   count_risingEdges = 0;      // Counter used to keep up with rising edge ticks for rpm calculation.
-float   rpm;                      // Variable used to convert electrical pulses to revolutions per minute for the fan.
+int   rpm;                        // Variable used to convert electrical pulses to revolutions per minute for the fan.
 
 
 // Initialization function
@@ -88,7 +88,7 @@ void loop() {
     sei();                                                        // Trigger the global interrupts
     delay(1000);                                                  // Delay loop by milisecond(s) interval to determine rpm.
     cli();                                                        // End the global interrupts
-    rpm = count_risingEdges/2.0;                                  // Each rising edge represents half of a rotation.
+    rpm = count_risingEdges/2;                                  // Each rising edge represents half of a rotation.
 
     // Write data to serial port for database thread to record
     Serial.println(tempValue);
